requirements.txt
streamlit
pandas
numpy
plotly
ibm-watson-machine-learning
python-dotenv
